+++ spinta/types/config.py
@@ -125,6 +125,8 @@ def load(context: Context, config: Config) -> Config:
 
     config.cache_control = rc.get("cache_control_header", default="")
 
+    config.http_strict_transport_security = rc.get("http_strict_transport_security", default="")
+
     if config.token_validation_keys_download_url and config.token_validation_key:
         raise ValueError(
             "token_validation_keys_download_url and token_validation_keys_download_url are mutually exclusive and can't be used together. Use one."